tips_analysis/requirements.txt
pandas==0.25.1
seaborn==0.9.0
matplotlib==3.1.1
scikit-learn==0.21.3
djangorestframework==3.14.0
